Fix mobile safe-area, ground/roof visuals, bridge+post collision
#39 (QA #6): touch.ts safeInset() read CSS vars --safe-bottom/--safe-right
that were never defined (parseFloat('')=NaN→0), so the joystick could overlap
the notch/home-bar. Defined both in index.html :root from env(safe-area-inset-*).

#42 (QA #11/#12): the ground shader blended zones over a 36u smoothstep band
that disagreed with the hard zoneAt() boundary — tightened the band (W 18→2) so
the ground colour closely tracks the gameplay zone. And the hollow-roof fade
triggered at the building's outer bounds (faded while you were still in the 2u
wall) — now gated on the interior (bounds ± WALL_T).

#43 (QA #18/#19): the bridge deck's and tunnel bore's outer ~0.8u were
unreachable (player r=0.8 vs a carve exactly the visual width) — expanded the
PASS_UNDER carves by the player radius so the FULL visual deck/bore is walkable
(verified: bridge edges at ±6.9 now walkable). And the gateway post collision
was 1.6×1.6 vs the 1.4 mesh — trimmed to ±0.7 (1.4). Suite 245/245.

+++ src/main.ts
@@ -214,7 +214,7 @@ async function start() {
       // warped zone radii — must mirror city.zoneAt (R0_BASE 200 ±34, R1_BASE 400 ±52)
       const warpLo = wLoA.x.mul(sin(th.add(wLoP.x))).add(wLoA.y.mul(sin(th.mul(2).add(wLoP.y)))).add(wLoA.z.mul(sin(th.mul(3).add(wLoP.z))));
       const warpHi = wHiA.x.mul(sin(th.add(wHiP.x))).add(wHiA.y.mul(sin(th.mul(2).add(wHiP.y)))).add(wHiA.z.mul(sin(th.mul(3).add(wHiP.z))));
-      const r0 = float(200).add(warpLo.mul(34)), r1 = float(400).add(warpHi.mul(52)), W = float(18);
+      const r0 = float(200).add(warpLo.mul(34)), r1 = float(400).add(warpHi.mul(52)), W = float(2); // tight band ≈ hard zoneAt() boundary (was 18 → 36u blend mismatched the gameplay zone)
       const sInner = smoothstep(r0.sub(W), r0.add(W), d), sOuter = smoothstep(r1.sub(W), r1.add(W), d);
       const fDown = oneMinus(sInner), fSub = sInner.sub(sOuter), fPark = sOuter;
       // shared mottle + crack network (mx_noise_float — identical on WebGPU + WebGL2, unlike worley)
